requirements.txt
streamlit>=1.33
openai>=1.40
pillow>=10.0
reportlab>=4.0
pytest>=8.0
